Reduce code duplication, component separation
