+++ jobs/Frontend/src/components/shared/Error.tsx
@@ -1,0 +1,3 @@
+export function NetworkErrorMessage() {
+    return <span>Failed to load data, please reload this page or came back later</span>;
+}
